Share cart totals and checkout controls with the agent
